requirements.txt
pydicom==2.3.1
opencv-python-headless==4.7.0.72
Pillow==9.5.0
numpy==1.24.3
